koruby_precise: node.def の算術 fast path が 32bit で切り詰まっていた
ruby.wasm と答えを突き合わせて発覚。while ループの総和が
31999996000000 ではなく 342161152 になっていた。

node.def の Fixnum 算術 fast path (__builtin_add_overflow 等) が intptr_t の
ままで、wasm32 では 32bit に切り詰まる。**node.def は生成コードの元なので、
先の .c/.h 一括置換 (716 箇所) から漏れていた** — 生成物 node_eval.c を見て
いれば気付けたが、ホストでは intptr_t == int64_t なので make test は
100523 のまま何も言わない。

あわせて printf 系の (long)FIX2LONG → (long long) + %lld (12 箇所)。
long は wasm32 では 32bit である。

これでベンチ 6 本 (fib/while/times/map/String/Hash) の答えが ruby.wasm と
一致した。

ホスト: make test 100523 (変化なし)、optcarrot AOT checksum 60838。
node.def を触ったので optcarrot は必須のゲート。

+++ sample/koruby_precise/builtins/array.c
@@ -182,7 +182,7 @@ static RESULT korb_m_ary_aref(CTX *c, VALUE *slots, VALUE_REF self, VALUE_SLICE 
         if (ev == KORB_NIL) { if (s > 0) { e = (korb_sword_t)n; e_incl = false; } else { e = 0; e_incl = true; } }
         else {
             e = FIX2LONG(ev); if (e < 0) e += n; e_incl = !excl;
-            if (UNLIKELY(e > (korb_sword_t)n)) return korb_raise(c, slots, KORB_E_RANGE, 0, "%ld out of range", (long)FIX2LONG(ev));
+            if (UNLIKELY(e > (korb_sword_t)n)) return korb_raise(c, slots, KORB_E_RANGE, 0, "%lld out of range", (long long)FIX2LONG(ev));
         }
         slots[0] = UNWRAP(korb_ary_new(c, slots, 8));       /* result (rooted) */
         VALUE_REF out = VALUE_REF_AT(&slots[0]);
